requirements.txt
pytube==10.9.2
Pillow
flask-restful
#pytest
# yt-dlp
# youtube-dlp
# streamlink
# streamlink_cli
